FileJournal: fix potential overflow in create()
CID 717016: Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN)At (1):
Potentially overflowing expression "g_conf->osd_max_write_size << 20" with type
"int" (32 bits, signed) is evaluated using 32-bit arithmetic before being used
in a context which expects an expression of type "int64_t" (64 bits, signed).
To avoid overflow, cast the left operand to "int64_t" before performing the
left shift.

+++ src/os/FileJournal.cc
@@ -427,7 +427,7 @@ int FileJournal::create()
     goto free_buf;
   }
 
-  needed_space = g_conf->osd_max_write_size << 20;
+  needed_space = ((int64_t)g_conf->osd_max_write_size) << 20;
   needed_space += (2 * sizeof(entry_header_t)) + get_top();
   if (header.max_size - header.start < needed_space) {
     derr << "FileJournal::create: OSD journal is not large enough to hold "
